bot/requirements.txt
aiohttp==3.7.4.post0
argh==0.26.2
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
dateparser==1.0.0
discord.py==1.7.3
Django==3.2.7
dj-database-url==0.5.0
flake8==3.9.2
fitbit==0.3.1
humanize==3.11.0
idna==3.2
kaleido==0.2.1
multidict==5.1.0
pandas==1.3.3
plotly==5.2.1
pycoingecko==2.2.0
psycopg2-binary==2.9.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.2
table2ascii==0.1.3
typing-extensions==3.10.0.2
urllib3==1.26.6
watchdog==2.1.5
yarl==1.6.3
yfinance==0.1.63